breakpoints & ranges -> into consts and update global storybook parameters
