+++ .storybook/preview.js
@@ -1,21 +1,67 @@
 import {addons} from '@storybook/addons'
 import {withThemeProvider, withSurroundingElements, toolbarTypes} from '../src/utils/story-helpers'
-import theme from './theme'
+import theme from '../src/theme-preval'
 
 export const globalTypes = toolbarTypes
 export const decorators = [withThemeProvider, withSurroundingElements]
 
+// TODO: Update the hard-coded values when the primitives are ready
+const PrimerViewports = {
+  xsmall: {
+    name: 'Xsmall',
+    styles: {
+      width: '320px',
+      height: '100%',
+    },
+  },
+  small: {
+    name: 'Small',
+    styles: {
+      width: theme.breakpoints[0],
+      height: '100%',
+    },
+  },
+  medium: {
+    name: 'Medium',
+    styles: {
+      width: theme.breakpoints[1],
+      height: '100%',
+    },
+  },
+  large: {
+    name: 'Large',
+    styles: {
+      width: theme.breakpoints[2],
+      height: '100%',
+    },
+  },
+  xlarge: {
+    name: 'Xlarge',
+    styles: {
+      width: theme.breakpoints[3],
+      height: '100%',
+    },
+  },
+  xxlarge: {
+    name: 'XXlarge',
+    styles: {
+      width: '1400px',
+      height: '100%',
+    },
+  },
+}
+
 addons.setConfig({
   // Some stories may set up keyboard event handlers, which can be interfered
   // with by these keyboard shortcuts.
-  enableShortcuts: false
+  enableShortcuts: false,
 })
 
 export const parameters = {
   actions: {argTypesRegex: '^on[A-Z].*'},
   html: {
     root: '#html-addon-root',
-    removeEmptyComments: true
+    removeEmptyComments: true,
   },
   options: {
     storySort: (a, b) => {
@@ -29,15 +75,15 @@ export const parameters = {
             [
               '*',
               // Within a set of stories, set the order to the following
-              ['*', 'Playground', /Playground$/, 'Features', 'Examples']
-            ]
-          ]
+              ['*', 'Playground', /Playground$/, 'Features', 'Examples'],
+            ],
+          ],
         ],
         'Behaviors',
         'Hooks',
         'Private components',
         'Deprecated components',
-        '*'
+        '*',
       ]
 
       /**
@@ -135,6 +181,11 @@ export const parameters = {
       }
 
       return compare(getHierarchy(a), getHierarchy(b))
-    }
-  }
+    },
+  },
+  viewport: {
+    viewports: {
+      ...PrimerViewports,
+    },
+  },
 }
