+++ src/Hidden/examples.stories.tsx
@@ -15,58 +15,8 @@ const meta: Meta = {
   args: {},
 }
 
-// See if there is an interest to take this into global params
-
-const PrimerViewports = {
-  xsmall: {
-    name: 'Xsmall',
-    styles: {
-      width: '320px',
-      height: '100%',
-    },
-  },
-  small: {
-    name: 'Small',
-    styles: {
-      width: '540px',
-      height: '100%',
-    },
-  },
-  medium: {
-    name: 'Medium',
-    styles: {
-      width: '768px',
-      height: '100%',
-    },
-  },
-  large: {
-    name: 'Large',
-    styles: {
-      width: '1012px',
-      height: '100%',
-    },
-  },
-  xlarge: {
-    name: 'Xlarge',
-    styles: {
-      width: '1280px',
-      height: '100%',
-    },
-  },
-  xxlarge: {
-    name: 'XXlarge',
-    styles: {
-      width: '1400px',
-      height: '100%',
-    },
-  },
-}
-
 const setViewportParamToNarrow = {
   viewport: {
-    viewports: {
-      ...PrimerViewports,
-    },
     defaultViewport: 'small',
   },
 }
